Mengimplementasikan fitur ubah data profil pengguna dan beberapa perbaikan tampilan

+++ lib/widgets/reservation_checkbox.dart
@@ -4,7 +4,7 @@ import 'package:pet_care_mobile_apps/styles/styles.dart';
 class ReservationCheckBox extends StatefulWidget {
   bool value;
   String serviceName;
-  double servicePrice;
+  int servicePrice;
   Function(bool?) onChanged;
 
   ReservationCheckBox({
